// src/app/model/customer.ts
export interface Customer {
    idCustomer?: any;
    firstNameCustomer: string;
    lastNameCustomer: string;
    cpfCustomer: any;
    birthdateCustomer: any;
    dateCreatedCustomer: any;
    monthlyIncomeCustomer: string;
    statusCustomer: boolean;
    emailCustomer: string;
    passwordCustomer: string;
}